refactor(giveback): split PersonalRoadmap into focused files
The roadmap was one ~870-line file. Split by responsibility, no behavior
change:
- givebackRoadmapTypes.ts: shared RoadmapLevel / ConnectorFill / RoadmapNode
- GivebackRoadmapRail.tsx: the Connector + RailToggle rail primitives
- GivebackRoadmapNode.tsx: the NodeRow marker/claim rendering
- GivebackPersonalRoadmap.tsx: the orchestrator (data shaping + layout)
